cdk/requirements.txt
attrs==21.4.0
aws-cdk.assets==1.155.0
aws-cdk.aws-acmpca==1.155.0
aws-cdk.aws-apigateway==1.155.0
aws-cdk.aws-apigatewayv2==1.155.0
aws-cdk.aws-apigatewayv2-integrations==1.155.0
aws-cdk.aws-applicationautoscaling==1.155.0
aws-cdk.aws-autoscaling==1.155.0
aws-cdk.aws-autoscaling-common==1.155.0
aws-cdk.aws-autoscaling-hooktargets==1.155.0
aws-cdk.aws-certificatemanager==1.155.0
aws-cdk.aws-cloudformation==1.155.0
aws-cdk.aws-cloudfront==1.155.0
aws-cdk.aws-cloudwatch==1.155.0
aws-cdk.aws-codebuild==1.155.0
aws-cdk.aws-codecommit==1.155.0
aws-cdk.aws-codeguruprofiler==1.155.0
aws-cdk.aws-codepipeline==1.155.0
aws-cdk.aws-codestarnotifications==1.155.0
aws-cdk.aws-cognito==1.155.0
aws-cdk.aws-ec2==1.155.0
aws-cdk.aws-ecr==1.155.0
aws-cdk.aws-ecr-assets==1.155.0
aws-cdk.aws-ecs==1.155.0
aws-cdk.aws-ecs-patterns==1.155.0
aws-cdk.aws-efs==1.155.0
aws-cdk.aws-elasticache==1.155.0
aws-cdk.aws-elasticloadbalancing==1.155.0
aws-cdk.aws-elasticloadbalancingv2==1.155.0
aws-cdk.aws-events==1.155.0
aws-cdk.aws-events-targets==1.155.0
aws-cdk.aws-globalaccelerator==1.155.0
aws-cdk.aws-iam==1.155.0
aws-cdk.aws-kinesis==1.155.0
aws-cdk.aws-kinesisfirehose==1.155.0
aws-cdk.aws-kms==1.155.0
aws-cdk.aws-lambda==1.155.0
aws-cdk.aws-logs==1.155.0
aws-cdk.aws-route53==1.155.0
aws-cdk.aws-route53-targets==1.155.0
aws-cdk.aws-s3==1.155.0
aws-cdk.aws-s3-assets==1.155.0
aws-cdk.aws-sam==1.155.0
aws-cdk.aws-secretsmanager==1.155.0
aws-cdk.aws-servicediscovery==1.155.0
aws-cdk.aws-signer==1.155.0
aws-cdk.aws-sns==1.155.0
aws-cdk.aws-sns-subscriptions==1.155.0
aws-cdk.aws-sqs==1.155.0
aws-cdk.aws-ssm==1.155.0
aws-cdk.aws-stepfunctions==1.155.0
aws-cdk.cloud-assembly-schema==1.155.0
aws-cdk.core==1.155.0
aws-cdk.custom-resources==1.155.0
aws-cdk.cx-api==1.155.0
aws-cdk.region-info==1.155.0
aws-cdk.pipelines==1.155.0
cattrs==22.1.0
constructs==3.3.283
exceptiongroup==1.0.0rc5
jsii==1.57.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.2.0
